app/requirements.txt
alembic==1.14.1 ; python_version >= "3.12" and python_version < "4.0"
amqp==5.3.1 ; python_version >= "3.12" and python_version < "4.0"
annotated-types==0.7.0 ; python_version >= "3.12" and python_version < "4.0"
anyio==4.8.0 ; python_version >= "3.12" and python_version < "4.0"
asyncpg==0.30.0 ; python_version >= "3.12" and python_version < "4.0"
bcrypt==4.0.1 ; python_version >= "3.12" and python_version < "4.0"
billiard==4.2.1 ; python_version >= "3.12" and python_version < "4.0"
celery==5.4.0 ; python_version >= "3.12" and python_version < "4.0"
certifi==2025.1.31 ; python_version >= "3.12" and python_version < "4.0"
cffi==1.17.1 ; python_version >= "3.12" and python_version < "4.0" and platform_python_implementation != "PyPy"
cfgv==3.4.0 ; python_version >= "3.12" and python_version < "4.0"
click-didyoumean==0.3.1 ; python_version >= "3.12" and python_version < "4.0"
click-plugins==1.1.1 ; python_version >= "3.12" and python_version < "4.0"
click-repl==0.3.0 ; python_version >= "3.12" and python_version < "4.0"
click==8.1.8 ; python_version >= "3.12" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.12" and python_version < "4.0" and (sys_platform == "win32" or platform_system == "Windows")
cryptography==44.0.0 ; python_version >= "3.12" and python_version < "4.0"
distlib==0.3.9 ; python_version >= "3.12" and python_version < "4.0"
dnspython==2.7.0 ; python_version >= "3.12" and python_version < "4.0"
ecdsa==0.19.0 ; python_version >= "3.12" and python_version < "4.0"
email-validator==2.2.0 ; python_version >= "3.12" and python_version < "4.0"
fastapi-cli[standard]==0.0.7 ; python_version >= "3.12" and python_version < "4.0"
fastapi[standard]==0.115.8 ; python_version >= "3.12" and python_version < "4.0"
filelock==3.17.0 ; python_version >= "3.12" and python_version < "4.0"
flower==2.0.1 ; python_version >= "3.12" and python_version < "4.0"
gevent==24.11.1 ; python_version >= "3.12" and python_version < "4.0"
greenlet==3.1.1 ; python_version >= "3.12" and platform_python_implementation == "CPython" and python_version < "4.0" or python_version >= "3.12" and python_version < "3.14" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32")
h11==0.14.0 ; python_version >= "3.12" and python_version < "4.0"
httpcore==1.0.7 ; python_version >= "3.12" and python_version < "4.0"
httptools==0.6.4 ; python_version >= "3.12" and python_version < "4.0"
httpx==0.28.1 ; python_version >= "3.12" and python_version < "4.0"
humanize==4.11.0 ; python_version >= "3.12" and python_version < "4.0"
identify==2.6.6 ; python_version >= "3.12" and python_version < "4.0"
idna==3.10 ; python_version >= "3.12" and python_version < "4.0"
itsdangerous==2.2.0 ; python_version >= "3.12" and python_version < "4.0"
jinja2==3.1.5 ; python_version >= "3.12" and python_version < "4.0"
kombu==5.4.2 ; python_version >= "3.12" and python_version < "4.0"
loguru==0.7.3 ; python_version >= "3.12" and python_version < "4.0"
mako==1.3.8 ; python_version >= "3.12" and python_version < "4.0"
markdown-it-py==3.0.0 ; python_version >= "3.12" and python_version < "4.0"
markupsafe==3.0.2 ; python_version >= "3.12" and python_version < "4.0"
mdurl==0.1.2 ; python_version >= "3.12" and python_version < "4.0"
nodeenv==1.9.1 ; python_version >= "3.12" and python_version < "4.0"
passlib==1.7.4 ; python_version >= "3.12" and python_version < "4.0"
platformdirs==4.3.6 ; python_version >= "3.12" and python_version < "4.0"
pre-commit==4.1.0 ; python_version >= "3.12" and python_version < "4.0"
prometheus-client==0.21.1 ; python_version >= "3.12" and python_version < "4.0"
prompt-toolkit==3.0.50 ; python_version >= "3.12" and python_version < "4.0"
pyasn1==0.6.1 ; python_version >= "3.12" and python_version < "4.0"
pycparser==2.22 ; python_version >= "3.12" and python_version < "4.0" and platform_python_implementation != "PyPy"
pydantic-core==2.27.2 ; python_version >= "3.12" and python_version < "4.0"
pydantic==2.10.6 ; python_version >= "3.12" and python_version < "4.0"
pygments==2.19.1 ; python_version >= "3.12" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.12" and python_version < "4.0"
python-dotenv==1.0.1 ; python_version >= "3.12" and python_version < "4.0"
python-jose[cryptography]==3.3.0 ; python_version >= "3.12" and python_version < "4.0"
python-multipart==0.0.20 ; python_version >= "3.12" and python_version < "4.0"
pytz==2025.1 ; python_version >= "3.12" and python_version < "4.0"
pyyaml==6.0.2 ; python_version >= "3.12" and python_version < "4.0"
redis==5.2.1 ; python_version >= "3.12" and python_version < "4.0"
rich-toolkit==0.13.2 ; python_version >= "3.12" and python_version < "4.0"
rich==13.9.4 ; python_version >= "3.12" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.12" and python_version < "4"
setuptools==75.8.0 ; python_version >= "3.12" and python_version < "4.0"
shellingham==1.5.4 ; python_version >= "3.12" and python_version < "4.0"
six==1.17.0 ; python_version >= "3.12" and python_version < "4.0"
slugify==0.0.1 ; python_version >= "3.12" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.12" and python_version < "4.0"
sqlalchemy==2.0.37 ; python_version >= "3.12" and python_version < "4.0"
starlette==0.45.3 ; python_version >= "3.12" and python_version < "4.0"
tornado==6.4.2 ; python_version >= "3.12" and python_version < "4.0"
typer==0.15.1 ; python_version >= "3.12" and python_version < "4.0"
typing-extensions==4.12.2 ; python_version >= "3.12" and python_version < "4.0"
tzdata==2025.1 ; python_version >= "3.12" and python_version < "4.0"
uvicorn[standard]==0.34.0 ; python_version >= "3.12" and python_version < "4.0"
uvloop==0.21.0 ; (sys_platform != "win32" and sys_platform != "cygwin") and platform_python_implementation != "PyPy" and python_version >= "3.12" and python_version < "4.0"
vine==5.1.0 ; python_version >= "3.12" and python_version < "4.0"
virtualenv==20.29.1 ; python_version >= "3.12" and python_version < "4.0"
watchfiles==1.0.4 ; python_version >= "3.12" and python_version < "4.0"
wcwidth==0.2.13 ; python_version >= "3.12" and python_version < "4.0"
websockets==14.2 ; python_version >= "3.12" and python_version < "4.0"
win32-setctime==1.2.0 ; python_version >= "3.12" and python_version < "4.0" and sys_platform == "win32"
zope-event==5.0 ; python_version >= "3.12" and python_version < "4.0"
zope-interface==7.2 ; python_version >= "3.12" and python_version < "4.0"
